Reorganize Home and its child components
