+++ frontend/src/components/elements/home/styles/Home.css
